Implement Tic Tac Toe game logic
A significant functionality update was implemented in the application. This includes the introduction of game logic into the App component in App.jsx. The update also includes the creation of GameBoard and Log components which display the game board and a log of game turns, respectively. Changes were also made to the Player component to highlight the active player. A deriveActivePlayer function was also introduced to determine the next player in the game.